Change frontend init code
Replaces inline JavaScript-initialization, because the websocket
keeps being re-initialized on page navigation. This is done by
removing the `getTurboStreamPageContribution()` function, but
providing a `getWebserviceUrl()` function instead.

The documentation now states that an init JS-file in assets should
be used instead.

Also update documentation to encourage use of webjar for frontend
Turbo-library.